Get_prn.md
Algorithm:
1) start
2) define function get_prn
3) create char variable 
4) accept input from user
5)print the same input
6) stop
Description:
This program can be used to get a roll number of required number of digits(in this case 11 digits). All this program does is defines a functions and in it a string ,which can be defined with whatever size we want to store. Then the input is asked from user and stored in the string and the string is later on printed on the screen.